An S4 object representing the SEIRDV.

This class represents the SEIRV model, showing how populations of susceptible, exposed, infectious and recovered individuals evolve over time. Vaccinated individuals are considered in their own compartment.

Usage

# S4 method for SEIRDV
initial_conditions(object)

# S4 method for SEIRDV
transmission_parameters(object)

# S4 method for SEIRDV
initial_conditions(object) <- value

# S4 method for SEIRDV
transmission_parameters(object) <- value

# S4 method for SEIRDV
intervention_parameters(object)

# S4 method for SEIRDV
intervention_parameters(object) <- value

# S4 method for SEIRDV
run(object, times = seq(0, 100, by = 1), solve_method = "lsoda")

Arguments

object

an object of the class SEIRDV

value

(list) list of intervention parameters: starts, stops and coverages.

times

(double) a sequence of time points at which the solution to the system of ODEs should be returned. Must be of the form seq(t_start, t_end, by=t_step). Default time series is seq(0, 100, by = 1).

solve_method

(string) a string of chosen numerical integration method for solving the ode system. Default is lsoda which is also the default for the ode function in the deSolve package used in this function.

Value

object of class SEIRDV with initial conditions assigned. object of class SEIRDV with transmission parameter values assigned. object of class SEIRDV with intervention parameters assigned. two dataframes: one with the time steps, age range, time series of S, E, I, R and V population fractions, and one with the time steps, age range, time series of incidences and deaths population fraction.

Methods (by generic)

initial_conditions: Retrieves initial conditions of SEIRV model.

transmission_parameters: Retrieves transmission parameters of SEIRV model.

initial_conditions<-: Setter method for initial conditions (S0, E0, I0, R0, V0, VR0) of the SEIRV model.

All initial conditions must sum up to 1. If the initial conditions provided to do not sum to 1, an error is thrown.

transmission_parameters<-: Set transmission parameters (beta, kappa, gamma, mu, nu, delta_V, delta_R) of the SEIRV model.

If the transmission parameters provided to are not 1-dimensional an error is thrown.

intervention_parameters: Retrieves intervention parameters of SEIRDV model.

intervention_parameters<-: Setter method for intervention parameters of the SEIRV model.

Intervention parameters have same size. A tanh function is used to smooth interventions during simulation. This class is designed for interventions which last several days at least and have several days between them; interventions involving rapid fluctuations may be distorted.

run: Solves ODEs of the SEIRDV specified in object for the time points specified in times and integration method specified in solve_method.

$$\frac{dS(t)}{dt} = - beta S(t) I(t) - nu Inter(t) S(t) + delta_V V(t) + delta_R R(t) + delta_VR VR(t)$$ $$\frac{dE(t)}{dt} = beta S(t) I(t) - kappa E(t)$$ $$\frac{dI(t)}{dt} = kappa E(t) - (gamma + mu) I(t)$$ $$\frac{dR(t)}{dt} = gamma I(t) - delta_R R(t) - nu Inter(t) R(t)$$ $$\frac{dV(t)}{dt} = nu Inter(t) S(t) - delta_V V(t)$$ $$\frac{dVR(t)}{dt} = nu Inter(t) R(t) - delta_VR VR(t)$$ $$\frac{dC(t)}{dt} = beta S(t) I(t)$$ $$\frac{dD(t)}{dt} = mu I(t)$$

where Inter(t) is the value at time t of the intervention protocol defined by the intervention parameters. This function relies on the package deSolve.

Slots

output_names

list of compartments name which are used by the model and incidence.

initial_condition_names

list of names of initial conditions (characters). Default is list("S0", "E0", "I0", R0", "V0", "VR0").

transmission_parameter_names

list of names of transmission parameters (characters). Default is list("beta", "kappa", "gamma", "mu", "nu", "delta_V", "delta_R", "delta_VR").

intervention_parameter_names

list of names of intervention parameters (characters). Default is list ("starts", "stops", "coverages")

initial_conditions

list of values for initial conditions (double).

transmission_parameters

list of values for transmission parameters (double).

intervention_parameters

list of values for intervention parameters (double).