agedemographics_by_country

A data frame describing the number of individuals (in thousands) per age group. Age groups span five years, starting with 0-4 years, 5-9 years, etc. with the oldest age group being 100+ years. Columns of the data frame are

Index number

Variant: type of data in the table row. Mostly estimates

Region, subregion, country or area: lists region to which data in the row applies.

Notes

Country code: three-letter code referring to country or region.

Type

Parent code

Reference data: year of count or estimate

0-4: count (in thousands) of individuals in the age group 0-4.

5-9: count (in thousands) of individuals in the age group 5-9.

10-14: count (in thousands) of individuals in the age group 10-14.

15-19: count (in thousands) of individuals in the age group 15-19.

20-24: count (in thousands) of individuals in the age group 20-24.

25-29: count (in thousands) of individuals in the age group 25-29.

30-34: count (in thousands) of individuals in the age group 30-34.

35-39: count (in thousands) of individuals in the age group 35-39.

40-44: count (in thousands) of individuals in the age group 40-44.

45-49: count (in thousands) of individuals in the age group 45-49.

50-54: count (in thousands) of individuals in the age group 50-54.

55-59: count (in thousands) of individuals in the age group 55-59.

60-64: count (in thousands) of individuals in the age group 60-64.

65-69: count (in thousands) of individuals in the age group 65-69.

70-74: count (in thousands) of individuals in the age group 70-74.

75-79: count (in thousands) of individuals in the age group 75-79.

80-84: count (in thousands) of individuals in the age group 80-84.

85-89: count (in thousands) of individuals in the age group 85-89.

90-94: count (in thousands) of individuals in the age group 90-94.

95-99: count (in thousands) of individuals in the age group 95-99.

100+: count (in thousands) of individuals in the age group 100+.

Usage

data(agedemographics_by_country)

Format

A dataframe with 29 variables and 18105 observations.

Source

UN Department of Economic and Social Affairs: Populations dynamics. World Population Prospects 2019